refactor(frontend): consolidate GRN hooks into useGRNWorkspace
